Avanzando con Anna y Facu la lista de Instrucciones

+++ cpu/src/atender_cpu.c
@@ -3,7 +3,7 @@
 void* atender_memoria()
 {
 	while (1) {
-		int cod_op = recibir_codOp(fd_memoria);
+		uint8_t cod_op = recibir_codOp(fd_memoria);
 		switch (cod_op) {
             case MENSAJE:
                 //
@@ -22,7 +22,7 @@ void* atender_memoria()
 void* atender_kernel_dis()
 {
 	while (1) {
-		int cod_op = recibir_codOp(fd_kernel_dis);
+		uint8_t cod_op = recibir_codOp(fd_kernel_dis);
 		switch (cod_op) {
             case MENSAJE:
                 //
@@ -41,7 +41,7 @@ void* atender_kernel_dis()
 void* atender_kernel_int()
 {
 	while (1) {
-		int cod_op = recibir_codOp(fd_kernel_int);
+		uint8_t cod_op = recibir_codOp(fd_kernel_int);
 		switch (cod_op) {
             case MENSAJE:
                 //
